A lot of things
I added a mode selector to edit liderhead and hitcircle separatly, so he can export hitcircle and slider head circle in the same time. For that, I've added 3 buttons: 'Hitcircle', 'Slider' and 'Numbers'. Now options aren't all displayed in the same time.

I've also fixed the bug with the coulours in the skin.ini because I put `Color1` instead of `Combo1`. It's now fixed.

Now I have to do the documentation too. I don't want to use the html because it's horrible, so I'll use the wiki of the GitHub, and maybe add a function which can read the Wiki and display this into the page.