Remove the help content
Because he was moved into the GitHub wiki. (I've also added the link btw).
I didn't removed the CSS in the case we can put the content of the Wiki directly in the website.
Also, I removed demo_mult_blend_mode.js, used in the help to show the multiplicative blend mode, now useless...